+++ src/pages/users/[uid]/lists/[listId]/index.page.tsx
@@ -1,11 +1,10 @@
-import { ParsedUrlQuery } from 'querystring';
 import { GetServerSideProps } from 'next';
 import React from 'react';
-import { List } from '../../../../../components/organisms/List';
-import { ListContent } from '../../../../../components/organisms/ListContent';
 import Layout from '../../../../../components/templates/layout';
 import { db } from '../../../../../libs/firebase';
 import { Params } from '../../../../../types/params';
+import { ListContent } from './ListContent';
+import { ListHeader } from './ListHeader';
 
 type Props = {
   listName: string;
@@ -19,7 +18,7 @@ const ListPage: React.FC<Props> = (props) => {
       image={`https://res.cloudinary.com/dhho8x7av/image/upload/l_text:Sawarabi%20Gothic_50_bold:${listName},co_rgb:333,w_800,c_fit/v1639747883/ogp_v2pkyb.png`}
     >
       <main className='flex flex-col bg-main-color'>
-        <List />
+        <ListHeader />
         <ListContent />
       </main>
     </Layout>
